+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## Changes
+- Bumped 1.21.7 release to 1.21.8.
+- Expanded allowed version ranges.
+
 ## Fixes
-- Fix Gradle thinking it was still 1.21.6
+- Add a workaround for Neoforge loading client resourcepacks on a dedicated server.